Changed tinestamp from int32 to int64

+++ services/swap/atom.go
@@ -15,5 +15,5 @@ type Atom interface {
 	Deserialize([]byte) error
 	GetFromAddress() ([]byte, error)
 	PriorityCode() uint32
-	RedeemedAt() (int32, error)
+	RedeemedAt() (int64, error)
 }
